fix : 게시글 thumbnail path 조회 에러 수정
## 프론트 전달 사항
- 반드시 `keeper_files/thumbnail/default/default_thumbnail_rectangle.jpg`, `keeper_files/thumbnail/default/default_thumbnail_square.jpg` 이 두 이미지를 넣어주셔야 합니다.

## 수정 사항
- 썸네일을 따로 지정하지 않는 게시글 entity는 새로 추가될 때 thumbnailEntity row에 id=1인 thumbnailEntity를 가짐.
   (Controller에 코드가 따로 설정된 게 없다면, 회원이나 도서 같은 다른 Entity도 똑같이 적용될 것)
- 기본 썸네일 이미지를 가지는 Entity를 삭제할 때, 썸네일 이미지는 삭제되지 않음 (기본 이미지이므로 삭제되면 안됨)
- 기본 썸네일 이미지에 대한 ThumbnailEntity의 경우, 여러 엔티티가 FK로 가질 수 있으므로 @ManyToOne 설정을 해야 함.
  - 현재는 PostingEntity에 대해서만 수정을 했으므로, 다른 것들도 수정해야 함 ㅠㅠㅠ!!

+++ src/main/java/keeper/project/homepage/util/service/ThumbnailService.java
@@ -24,8 +24,17 @@ public class ThumbnailService {
 
   private final static String THUMBNAIL_FORMAT = "jpg";
 
-  private final String relDirPath = "keeper_files" + File.separator + "thumbnail";
-  private final String defaultImageName = "thumb_default.jpg"; // thumbnail delete시 파일명 비교
+  private final static String relDirPath = "keeper_files" + File.separator + "thumbnail";
+  private final static String defaultImageName = "thumb_default.jpg"; // thumbnail delete시 파일명 비교
+  // DB에는 디렉토리 구분자가 "/"
+  public final static String DEFAULT_RELATIVE_DIR_PATH = relDirPath + File.separator + "default";
+
+  // TODO : file과 thumbnail의 default id, name은 같지만, 그래도 각 service에 분리하기
+  // TODO : 앞으로 default 더 생길 것이므로, class나 enum으로 정리하기
+  public final static String DEFAULT_RECTAGLE_NAME = "default_thumbnail_rectangle.jpg";
+  public final static String DEFAULT_SQUARE_NAME = "default_thumbnail_square.jpg";
+  public final static Long DEFAULT_RECTANGLE_ID = 1L;
+  public final static Long DEFAULT_SQUARE_ID = 2L;
 
   private final ImageFormatChecking imageFormatChecking;
   private final ThumbnailRepository thumbnailRepository;
@@ -64,12 +73,22 @@ public class ThumbnailService {
     FileEntity fileEntity = null;
     String fileName = "";
     if (multipartFile == null || multipartFile.isEmpty()) {
-      fileName = this.defaultImageName;
-      File defaultFile = new File(
-          FileService.fileRelDirPath + File.separator + FileService.defaultImageFileName);
-      fileEntity = fileService.saveFileEntity(
-          defaultFile, FileService.fileRelDirPath, ipAddress, FileService.defaultImageFileName,
-          null);
+      FileEntity defaultFile = null;
+      ThumbnailEntity defaultThumbnail = null;
+      if (size.equals(ThumbnailSize.LARGE)) { // -> rectangle
+        defaultFile = fileService.findFileEntityById(DEFAULT_RECTANGLE_ID);
+        defaultThumbnail = findById(DEFAULT_RECTANGLE_ID);
+      } else { // -> square
+        defaultFile = fileService.findFileEntityById(DEFAULT_SQUARE_ID);
+        defaultThumbnail = findById(DEFAULT_SQUARE_ID);
+      }
+      return defaultThumbnail;
+//      fileName = this.defaultImageName;
+//      File defaultFile = new File(
+//          FileService.fileRelDirPath + File.separator + FileService.defaultImageFileName);
+//      fileEntity = fileService.saveFileEntity(
+//          defaultFile, FileService.fileRelDirPath, ipAddress, FileService.defaultImageFileName,
+//          null);
     } else {
       MultipartFileWrapper multipartFileWrapper = new MultipartFileWrapper(multipartFile);
       try {
@@ -93,12 +112,12 @@ public class ThumbnailService {
         // 업로드 임시 파일 삭제
         multipartFileWrapper.transferFinish();
       }
+      return thumbnailRepository.save(
+          ThumbnailEntity.builder()
+              .path(this.relDirPath + File.separator + fileName)
+              .file(fileEntity)
+              .build());
     }
-    return thumbnailRepository.save(
-        ThumbnailEntity.builder()
-            .path(this.relDirPath + File.separator + fileName)
-            .file(fileEntity)
-            .build());
   }
 
   public ThumbnailEntity findById(Long findId) {
@@ -107,16 +126,17 @@ public class ThumbnailService {
   }
 
   public void deleteById(Long deleteId) {
-    ThumbnailEntity deleted = findById(deleteId);
     // 기본 썸네일이면 삭제 X
-    if (deleted.getPath().equals(relDirPath + File.separator + defaultImageName) == false) {
-      File thumbnailFile = new File(
-          System.getProperty("user.dir") + File.separator + deleted.getPath());
-      if (thumbnailFile.exists() == false) {
-        throw new CustomFileNotFoundException();
-      } else if (thumbnailFile.delete() == false) {
-        throw new CustomFileDeleteFailedException();
-      }
+    if (deleteId == DEFAULT_RECTANGLE_ID || deleteId == DEFAULT_SQUARE_ID) {
+      return;
+    }
+    ThumbnailEntity deleted = findById(deleteId);
+    File thumbnailFile = new File(
+        System.getProperty("user.dir") + File.separator + deleted.getPath());
+    if (thumbnailFile.exists() == false) {
+      throw new CustomFileNotFoundException();
+    } else if (thumbnailFile.delete() == false) {
+      throw new CustomFileDeleteFailedException();
     }
     thumbnailRepository.deleteById(deleteId);
   }
